+++ frontend/src/components/Dashboard/DashboardProfile.jsx
@@ -1,0 +1,7 @@
+export default function LandingHome() {
+    return (
+        <div className="container mt-5">
+            <h>Profile</h>>
+        </div>
+    );
+}